fix(android): harden hierarchy parser and tighten tests
